requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.59.6
httpx==0.27.2
yt-dlp==2023.12.30
python-multipart==0.0.6
pydantic-settings==2.1.0